feat(patientinfodashboard): Enables fetching patients
This is before manually merging things with the SCRUM-44/AdminTable branch

Adds functionality to fetch patient data and display it in the Patient Info Dashboard.

This introduces a new API endpoint to retrieve all patients and integrates it into the dashboard. It also adds a basic delete functionality to remove patients.